Deletd unnecessary files.
