--- lib/app_screens/details/details.dart
@@ -1,12 +1,0 @@
-import 'package:flutter/material.dart';
-import 'details_attributes/DetailsBody.dart';
-
-// ignore: non_constant_identifier_names
-void detail_main(){
-  runApp(
-    MaterialApp(
-      title: "My Flutter App",
-      home: DetailPage(),
-    ),
-  );
-}
