--- lib/app_screens/discover/discover.dart
@@ -1,12 +1,0 @@
-import 'package:flutter/material.dart';
-import 'discover_attributes/DiscoverBody.dart';
-
-// ignore: non_constant_identifier_names
-void discover_main(){
-  runApp(
-    MaterialApp(
-      title: "My Flutter App",
-      home: DiscoverPage(),
-    )
-  );
-}
