+++ lib/app_screens/main/main_attributes/BodyAttributes.dart
@@ -1,7 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_app/app_screens/details/details.dart';
-import 'package:flutter_app/app_screens/details/details_attributes/DetailsBody.dart';
-import 'package:flutter_app/app_screens/main/Body.dart';
 import 'package:flutter_app/shared/helpers/icomoon.dart';
 import 'package:flutter_app/widgets/expanded/divider.dart';
 import 'package:flutter_app/widgets/expanded/widgets_attribute/Main/MainLists.dart';
