+++ lib/app_screens/myfavorites/MyFavorites.dart
@@ -1,13 +1,29 @@
 import 'package:flutter/material.dart';
-import 'myfavorites_attribute/MyFavoritesBody.dart';
+import 'package:flutter_app/app_screens/settings/dialog/Dialog.dart';
+import 'package:flutter_app/shared/helpers/icomoon.dart';
 
-// ignore: non_constant_identifier_names
-void MyFavorite_Main(){
-  runApp(
-    MaterialApp(
-      title: "My Flutter App",
-      home: MyFavoritesPage(),
-    ),
-  );
+class MyFavoritesPage extends StatelessWidget{
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        backgroundColor: Colors.black87,
+        title: Text("MY FAVORITES"),
+        centerTitle: true,
+      ),
+      body: MyFavoritesPageBody(),
+    );
+  }
 }
 
+class MyFavoritesPageBody extends StatelessWidget{
+  @override
+  Widget build(BuildContext context) {
+    return Center(
+      child: Container(
+
+      ),
+    );
+  }
+
+}
