--- lib/app_screens/myfavorites/myfavorites_attribute/MyFavoritesBody.dart
@@ -1,29 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_app/app_screens/settings/dialog/Dialog.dart';
-import 'package:flutter_app/shared/helpers/icomoon.dart';
-
-class MyFavoritesPage extends StatelessWidget{
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        backgroundColor: Colors.black87,
-        title: Text("MY FAVORITES"),
-        centerTitle: true,
-      ),
-      body: MyFavoritesPageBody(),
-    );
-  }
-}
-
-class MyFavoritesPageBody extends StatelessWidget{
-  @override
-  Widget build(BuildContext context) {
-    return Center(
-      child: Container(
-
-      ),
-    );
-  }
-
-}
